LayoutTests/fast/css/style-scoped/scope-pseudo.html

<!DOCTYPE html>
<html>
<head>
    <script type="text/javascript">
        function log(msg)
        {
            document.getElementById('console').appendChild(document.createTextNode(msg + '\n'));
        }

        function test(id)
        {
            var elem = document.getElementById(id);
            log(id + ': ' + document.defaultView.getComputedStyle(elem, null).getPropertyValue('color'));
        }

        function runTests()
        {
            if (window.testRunner)
                testRunner.dumpAsText();

            log('--- COMPUTED STYLES ---');

            log('<html> background: ' + document.defaultView.getComputedStyle(document.documentElement, null).getPropertyValue('background-color'));
            test('outer');
            test('sibling1');
            test('scope1');
            test('inner1d');
            test('inner1s');
            test('sibling2');
            test('scope2');
            test('inner2d');
            test('inner2s');
            test('sibling3');

            log('--- FINISHED ---');
        }
    </script>
    <style type="text/css">
        body { color: black; }
        :scope { background-color: yellow; }
    </style>
</head>
<body onload="runTests();">
    <p>Test :scope pseudo-class with &lt;style scoped&gt;</p>
    <div id="outer">
        <div id="sibling1">Text</div>
        <div id="scope1">
            <style type="text/css" scoped>
                div { color: yellow; }
                :scope { color: red; }
                div:scope { color: green; }
            </style>
            <div id="inner1d">Text</div>
            <span id="inner1s">Text<span>
        </div>
        <div id="sibling2">Text</div>
        <div id="scope2">
            <style type="text/css" scoped>
                :scope div { color: blue; }
            </style>
            <div id="inner2d">Text</div>
            <span id="inner2s">Text<span>
        </div>
        <div id="sibling3">Text</div>
    </div>
    <pre id="console"></pre>
</body>
</html>
