LayoutTests/fast/css/style-scoped/scope-pseudo-expected.txt

Test :scope pseudo-class with <style scoped>

Text
Text
Text
Text
Text
Text
Text
--- COMPUTED STYLES ---
<html> background: rgb(255, 255, 0)
outer: rgb(0, 0, 0)
sibling1: rgb(0, 0, 0)
scope1: rgb(0, 128, 0)
inner1d: rgb(255, 255, 0)
inner1s: rgb(0, 128, 0)
sibling2: rgb(0, 0, 0)
scope2: rgb(0, 0, 0)
inner2d: rgb(0, 0, 255)
inner2s: rgb(0, 0, 0)
sibling3: rgb(0, 0, 0)
--- FINISHED ---
